requirements.txt
streamlit>=1.28.0
google-genai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
statsmodels>=0.14.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0
pyngrok
#pydeck>=0.8.0
#streamlit-option-menu>=0.3.0
